backend/requirements.txt
# =============================================================================
# Jira QA AI Generator - Python Dependencies
# =============================================================================

# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
slowapi>=0.1.9
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
cryptography==42.0.2

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# LLM SDKs
google-generativeai==0.4.0
anthropic==0.18.1
openai==1.12.0

# Jira Integration
jira==3.6.0
atlassian-python-api==3.41.11

# Validation & Parsing
email-validator==2.1.0.post1
python-dateutil==2.8.2

# Logging & Monitoring
loguru==0.7.2
sentry-sdk[fastapi]==1.40.0
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0
factory-boy==3.3.0
faker==22.6.0

# Development Tools
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# Utilities
tenacity==8.2.3
python-dotenv==1.0.1
pyyaml==6.0.1
jinja2==3.1.3
